Finish css exercises 1-6
I initially forgot to fork this when I cloned it, I removed the origin
and changed it to the fork but I can't be sure that I did it right. If
anyone from project odin is seeing this please message me on github to
yell at me so I know not to push this repository again.